+++ CHANGELOG.md
@@ -42,6 +42,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 * [131](https://github.com/babylonlabs-io/babylon/pull/131) Add new staking
 events
 
+### Misc Improvements
+
+* [#113](https://github.com/babylonlabs-io/babylon/pull/113) Add multibuild binary
+for upgrade handler `testnet` and `mainnet`.
+
 ## v0.11.0
 
 ### State Machine Breaking